// src/cashier_frontend_new/src/modules/fee/type.ts
/**
 * Enumeration of fee types
 */
export class FeeType {
  static readonly NETWORK_FEE = new FeeType("NETWORK_FEE");
  static readonly CREATE_LINK_FEE = new FeeType("CREATE_LINK_FEE");

  private constructor(public readonly type: string) {}
}

/**
 * Fee item representation
 */
export interface FeeItem {
  feeType: FeeType;

  symbol: string;
  price?: number;
  /** in formmated string */
  amount: string;
  /** in formmated string */
  usdValueStr?: string;
  /** number usd value */
  usdValue?: number;
}